+++ package.json
@@ -5,6 +5,7 @@
   },
   "dependencies": {
     "aws-sdk": "^2.1233.0",
+    "dotenv": "^16.0.3",
     "jest": "^29.1.2",
     "uuid": "^9.0.0"
   }
